LayoutTests/imported/w3c/web-platform-tests/html/browsers/origin/origin-keyed-agent-clusters/popups/opener-no-openee-yes-same.sub.https-expected.txt


PASS message event must occur
PASS setting document.domain must give sync access
FAIL opener: originAgentCluster must equal false assert_equals: expected (boolean) false but got (undefined) undefined
FAIL openee: originAgentCluster must equal false assert_equals: expected (boolean) false but got (undefined) undefined
